change: move cli items to their own app folder
Moved all files related to the cli to their own folder under apps/cli so
that all tests and files related live in the correct directory
