+++ pytest.ini
@@ -1,5 +1,10 @@
 [pytest]
-testpaths = tests
+testpaths =
+    tests
+    apps/cli/tests
+pythonpath =
+    apps/cli
+    apps
 python_files = test_*.py
 python_classes = Test*
 python_functions = test_*
