+++ tools/install.sh
@@ -324,8 +324,9 @@ main() {
       # in frontend/ first if you changed any frontend code.
       local_src="$(cd "$FABRICATOR_LOCAL_SRC" && pwd)"
       info "Local source mode: staging from $local_src (skipping release download)"
-      mkdir -p "$APP_SRC_DIR/frontend"
+      mkdir -p "$APP_SRC_DIR/frontend" "$APP_SRC_DIR/apps"
       rsync -a --exclude='__pycache__/' --exclude='*.pyc' "$local_src/backend/" "$APP_SRC_DIR/backend/"
+      rsync -a --exclude='__pycache__/' --exclude='*.pyc' "$local_src/apps/cli/" "$APP_SRC_DIR/apps/cli/"
       rsync -a --exclude='__pycache__/' --exclude='*.pyc' "$local_src/tools/" "$APP_SRC_DIR/tools/"
       cp "$local_src/run.py" "$local_src/requirements.txt" "$local_src/pyproject.toml" "$APP_SRC_DIR/"
       if [[ -d "$local_src/frontend/dist" ]]; then
